pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-eval-framework"
version = "0.1.0"
description = "Automation-first framework for multi-model LLM evaluation."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Bhargob Deka" }]
dependencies = [
  "pydantic>=2.8.0",
  "PyYAML>=6.0.1",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "groq>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]

[project.scripts]
llm-eval = "llm_eval.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
